filestore: sync object_map on _set_replay_guard()
We need to sync the object_map too.  We can _almost_ check to see if there
are keys for the object and only do it then, except that they may have
existed previously and then been deleted.

So, always sync.  leveldb is reasonably nice about this... it should just
be another fsync.

Signed-off-by: Sage Weil <sage@newdream.net>