add smoke suite
This could probably be collapsed into a bunch of singleton tasks to make
it simpler to track how many actual jobs result, but it was simpler to
make it a subset of regression.  And probably that'll be easier to maintain
moving forward.

Tried to avoid any jobs that took more than 10 minutes (tho there are a few
in here).  Kept both valgrind and lockdep jobs, and dropped many of those
from the basic collection (esp api tests).

We'll see how long this takes on plana and adjust up/down from there,
depending on how long we want to wait for it.