ceph_mon: fix fsid parsing.
fsid is a field in the CephContext _conf structure and is parsed by
the standard options parsing library before it gets to the ceph_mon
custom parsing.
Instead do the standard parsing, and check that member directly
to decide if we want to (over)write the monmap's fsid.

Fixes one part of #2221.

Signed-off-by: Greg Farnum <gregory.farnum@dreamhost.com>