filestore: simple failure injections via --filestore-kill-at <n>
This will make filestore suicide (_exit(1)) on the n'th potential failure
call site.  We can potentially fail:

     - before a transaction
     - between each op
     - at the end

Additionally, we instrument the guards:

     - before/after/inside _set_replay_guard
     - between significant steps of callers of _set_replay_guard

All instrumentation points are inside _do_transactions(), so if everything
is done in a single sequencer (or from a single thread) the failure
point is deterministic.

That said, use an atomic so we will still reliably fail (at some point)
when there are multiple filestore threads in action.

Signed-off-by: Sage Weil <sage.weil@dreamhost.com>