qa: add run_xfsests.sh script
Add a script that runs xfstests over a pair of devices that are
specified using command line arguments.  The tests are run using
a specified filesystem type (xfs, ext4, or btrfs).

A default set of tests is run if none is specified on the command
line.  Normally there's an "auto" group used for this purpose, but
for now I've laid out a (large) subset of them that I know pass on
rbd devices.  These can be updated as we find they work reliably.

Signed-off-by: Alex Elder <elder@dreamhost.com>