msgr: Document remaining SimpleMessenger functions and data members
This doesn't handle the inner classes (Pipe, mostly) yet. It also
raises questions about whether these functions should be public, or if
we should declare all the inner classes friends.
After talking it over, I think we want them to be public, since
the Messenger interface is the important abstraction. But that
will require reworking some stuff anyway to make it consistent. ;)

Signed-off-by: Greg Farnum <gregory.farnum@dreamhost.com>