osd: fix dirty_info check for advance/activate paths
Previously we would check and write dirty_info *without the pg lock* after
doing the advance and activate map calls.  This was unlikely to race with
anything because the queues were drained, but definitely not right.

Instead, do the write in activate_map, or explicitly if activate_map is
not called (so that we record our progress after handling maps when we are
not up).

Signed-off-by: Sage Weil <sage.weil@dreamhost.com>