osd: dirty_info if history.merge updated anything
In proc_replica_info and proc_primary_info, we may or may not update
the pg_info_t.  If we do, set dirty_info, so that it will be recorded.
Same goes for when the primary pushes out updated stats to us.

Also, do not write a purged_snaps() update directory; rely on the caller
to write out dirty info.

Signed-off-by: Sage Weil <sage.weil@dreamhost.com>