crush: add check_item_loc
The check_item_loc() method will take an item and position and tell you if
it matches the items current location.  The matching is identical to that
used for insert_item, in that a specific location constraint match means
success, even if a less specific one does not match (e.g., rack=wrongrack,
host=correcthost will return true).

Signed-off-by: Sage Weil <sage@newdream.net>