msgr: If the Connection has a bad Pipe*, replace it with a good one.
This is a little weak, since it doesn't fill in the Connection's data
when we initially create the new Pipe. But doing that would require
passing more state around that mucks up the interfaces a bit, so I
don't think it's worthwhile. This also makes it clear that this
set of functions is not responsible for doing things like reset
notifications -- those need to be handled through the normal channels
(I reviewed them and as best I can tell, they are.)

Signed-off-by: Greg Farnum <greg@inktank.com>