msgr: pass a Connection instead of a Pipe through submit_message.
This lets us maintain a user's existing Connections even if the underlying
Pipe object disappears (ie, it's lossy and they don't drop the Connection
object).

This is a little distressing for me because it's again emphasizing the
Connection over the Pipe, but that seems to be an ongoing process.
Maybe the next Messenger implementation will be able to do away with
the (fairly messy) distinction between them.

Signed-off-by: Greg Farnum <greg@inktank.com>