FileStore,DBObjectMap: add SequencerPosition argument to ObjectMap
Previously, sequences like:

1. touch (c1, a)
2. link (c1, c2, a)
3. rm (c1, a)
4. setattr (c2, a)
5. clone (c2, a, b)

could result in the omap entries for a being removed once ops 2-3 are
replayed.  Calls to ObjectMap::sync will include a sequencer posotion
and an hobject_t to mark.  Updates to the object map will now also check
the SequencerPosition entry on the map header preventing replay of
earlier ops.

Signed-off-by: Samuel Just <sam.just@inktank.com>