msg: make clear_pipe work only on a given Pipe, rather than the current one.
This way old Pipes that have been replaced can't clear the new Pipe
out of a Connection's link.
We might attempt to instead sever the link between CLOSED Pipes and
their Connections more completely (eg, when the Connection gets a
new Pipe), but that will require more work to handle all the
cases, and this works for now.

Signed-off-by: Greg Farnum <greg@inktank.com>