OSD: do not convert an entire collection in one transaction
Previously, we atomically moved the collection out of the way, created a
new collection, moved the contents of the old collection to the new
collcetion and removed the old collection.  For large collections, this
could result in unacceptably long transactions.  Now, we create a temp
collection, link all objects into it, atomically swap them, and then
remove the old one.

Signed-off-by: Samuel Just <sam.just@inktank.com>