keyserver: also authenticate against mon keyring
If we don't have a secret, also check in the extra_secrets keyring.

This means we can also authenticate as any users that appear in the mon
keyring, and get the caps defined there.  This lets us bootstrap the
client.admin key with mon. key, provided mon 'allow *' caps appear in the
mon keyring.

Signed-off-by: Sage Weil <sage@inktank.com>