crush: make magic numbers tunable
We have three magic numbers in crush_choose that are now tunable.  The
first two control the local retry behavior, including fallback to a
permutation.  The last is the total map descent attempts.

We can avoid a drastic incompatibility by making these tunable and encoded
in the map.  That means users can enable/disable local retry, for example,
without changing the code.  As long as the clients understand the tunables,
they can be adjusted.

This patch doesn't address the compatibility and feature bit issue.  We may
want to roll that into a larger revision with more drastic changes, once
we know what those changes will look like.  However, a careful user can
use the new code and modify the behavior.

Signed-off-by: Sage Weil <sage@inktank.com>