osdcap: move semantics of cap check into OSDCap class
The class needs to understand the semantics of possible actions in enough
detail to apply the capabilities it describes, so it makes sense to pass
all of that information into a check method here.  That is preferable to
exposing the internal structure of the capability itself to an external
function to apply the same logic.

Specifically, this fixes the problem where we return the first matching
OSDCapSpec, not a union of all matching specs, and thus fixes a problem
where applying a cap like 'allow r, allow pool foo rwx' to a write
operation in pool foo.

Signed-off-by: Sage Weil <sage@inktank.com>