librbd: add indirection between name and header object
This adds a new rbd_id.$name object that contains the id of an image.
Rename the header in new-format images to  rbd_header.$id, so that
it doesn't need to change when an image is renamed - the rbd_id.$name
object is copied to a new object instead.

With layering, this becomes more important because many child images
could be depending on the header object of a parent image.

This is another change to the new format, but since no one uses it yet,
we don't need to worry about backwards compatibility.

Signed-off-by: Josh Durgin <josh.durgin@inktank.com>