cls_rbd: add methods for dealing with rbd_directory objects
These hide the structure of the directory from clients, and provide
mappings from id to name and name to id. These mappings aren't needed
for old format images, but with layering, they're useful to show
the user e.g. the name of an image's parent instead of its id.

This also makes the directory more efficient for new-style images, since
it uses omap instead of tmap.

Signed-off-by: Josh Durgin <josh.durgin@inktank.com>